sympy/core: Basic and Expr __eq__ are merged
